drag-drop/pyproject.toml
[project]
name = "drag-drop"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.10.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "wikipedia>=1.4.0",
]
